+++ application/ui/src/features/annotator/labels/labels-editor/labels-editor.module.scss
@@ -1,4 +1,4 @@
-.popoverContent {
+.editorContent {
     padding: var(--spectrum-global-dimension-size-150);
     min-width: var(--spectrum-global-dimension-size-6000);
 }
